+++ back-end/src/database/models/sales.js
@@ -7,7 +7,7 @@ const Sale = (sequelize, DataTypes) => {
     seller_id: {
       type: DataTypes.INTEGER, foreignKey: true,
     },
-    total_price: DataTypes.INTEGER,
+    total_price: DataTypes.DECIMAL,
     delivery_address: DataTypes.STRING,
     delivery_number: DataTypes.STRING,
     saleDate: {
